fix(backend): clear all 123 mypy errors and fix the real bugs behind them
`mypy app --ignore-missing-imports` (the lint-backend gate) now reports
"Success: no issues found in 210 source files", down from 123 errors
across 33 files. Every fix targets the root cause rather than silencing
the check.

Real bugs found and fixed along the way:

- models/relationship.py: RecipientSharedMemory never inherited
  TimestampMixin, so ordering by created_at raised AttributeError at
  runtime even though migration 022 creates the column.
- services/pricing, recommendations: `project.metadata` / `template.metadata`
  resolve to SQLAlchemy's MetaData object, not the JSONB column aliased as
  "metadata" -> `metadata_`. These raised AttributeError on every call.
- services/gallery: Generation has no `user_id`; ownership now resolves via
  requested_by_user_id with a project-owner fallback (same pattern as
  pipeline_tasks.py).
- api/v1/admin.py: GenerationJob has no `retry_count`; it is `attempts`.
- api/v1/lab.py: AdminUser has no `email`; resolve it from the linked User.
- api/v1/feedback.py: `feedback = repo.db.add(...)` assigned None and then
  serialised it; now calls the repository's create().
- services/payments: `Model.__table__.update()` -> SQLAlchemy 2.0 `update()`,
  and the RETURNING row is unwrapped instead of assigned as the entity.
- services/analytics/funnel.py: invalid `func.cast(str, func.interval())`;
  the cutoff is now computed in Python with a timedelta.
- services/prompt_compiler: get_by_id was called without the required
  owner_user_id, and required_variables read a nonexistent metadata_ column.
- services/ab_test: ABTestCreate lacked start_date/end_date that create_test
  already validated; the variant fallback returned an ORM row where a
  ABTestVariantCreate was declared.
- models/ab_test: traffic_allocation annotated float on an Integer column.

Typing corrections: the circuit_breaker decorator genuinely returns None on
an open circuit (script_generation already handles `if result is None`), so
its signature is now Awaitable[T | None] and the two decorated providers plus
their base classes were widened to match. `minio` is a declared dependency,
so the ImportError fallback that mypy rejected was dead code.

Also drops the duplicated password/name fields in AdminSetupRequest and
renames RecommendationService.list -> list_recommendations (the method
shadowed the builtin and corrupted `list[str]` annotations in the same class).

Verified: ruff check . and ruff check app tests both clean; 167 unit + 4
integration tests pass.

+++ backend/app/models/ab_test.py
@@ -17,7 +17,7 @@ class ABTest(Base, UUIDPrimaryKeyMixin, TimestampMixin):
     status: Mapped[str] = mapped_column(String(20), nullable=False, default="draft")
     start_date: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     end_date: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
-    traffic_allocation: Mapped[float] = mapped_column(Integer, nullable=False, default=100)
+    traffic_allocation: Mapped[int] = mapped_column(Integer, nullable=False, default=100)
     variant_a_id: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True))
     variant_b_id: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True))
     results_lock: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
